Revert "Refactor int computer"
